Use build cache for full image build
Build cache was not being used on the final image build, as the
CACHE_FROM variable is empty during that function call. This is caused
by the steps being built in a different sub shell, and as such any
variables set during the step builds will be wiped afterwards.

Address this issue by outputting and loading the build cache to a file