Add basic commit ref sanitization for tagging
Docker images don't support as large variety of characters in their tag names as git,
and as such we will run into issues attempting to tag everything the same way
as they are in git.

Take first steps in the sanitization effort by replacing commonly used invalid characters
with a dash.

This is not a complete solution, and a more future proof method should be thought of.

No refs