Revert reversing build order of Dockerfile stages
It turns out that caching did not work as we though and when using
multi-stage builds not all stages are cached if the final stage is built
first. The final image will not contain all of the cached
layers when running with DIND, instead only the last layers
will be present. This reverts back to the way v1 did building
by building from top to bottom.

Future thoughts: Use Kaniko