Refactor out Project and Database handling to their own classes
This is the first step into making everything a bit less coupled. There are
still things that needs to be done in order to get everything in order but
this is the start. The next step is to also split out the deployment, however
some of the things found in the Kubernetes class was already move to
the Project class, and will most likely end up the end future Deployment
class.

The update also makes way for allowing multi-project deployments
with project dependencies. This is solved by allowing for passing dependency
projects as env vars to the pipeline at this point. In the future the syntax
for this might change and will allow for better variable control for dependency
projects. The first solution is to read env vars which has been prefixed with
the name of the project (in uppercase without any hyphens).