Update pullPolicy to `Always` to make the risk for using stale images smaller
Kubernetes should have a cache that is uses anyways so old cache layers should
be present for a while on each node. There will for sure be a bit more traffic generated
by this, but we will try it out and if it turns out to be a too costly of an experiment we
will have to make the flag either dynamic or revert back.