Add service support
This adds services to replace the "hard-to-maintain" database deployments used before
and adds support for easily add support for new services such as for caching or message queues.

As part of this refactor, all previous automatic database handling has been removed and with that
the users of the pipeline will now need to specify which services to use without environment variables.

Note that this is the first version of the services and is therefor subject to change in some regards,
for instance the following issues resides:

1. Only a single env var can be specified as parameter for artifacts. Other artifacts can be added by
    the services themselves, but can at this point not be modified by the CI user.
2. PostgeSQL and MySQL does not have feature parity when it comes to multi-database setups