Always create file secrets
Before they were only created if there were actual data. This turned out to be problematic
because the current deployment is expecting there to be file secrets as the file secret name
always gets populated.

This is not an optimal solution and a proper solution should be to refactor the project class to
make sure that a name is not set if there is no secrets.