Add possibility to configure probe delay
At this stage we handle both of the probes the same way to keep things simple.
We could at a later point split this up to be more configurable. There are also several
other configurations that could be added regarding the probes, but as we don't have
a use case for them right now, I restrained myself from adding them.