Optimize helm timeout
The previous timeout was just based on the delay plus a buffer, the new one is based multiple variables
for a more exact timeout.

The first line is to make sure that all instances are actually up. If there is more than one replica, the N+1
replica does not start before the first one has successfully been set up. Therefor, if there are more than one
replica, we multiply the startup time by two.
Second line is so that we do not time out before Kubernetes has finished all probe checks.
Lastly we add a buffer of 2 minutes just to be sure that there is some breathing room.

Example:
3 replicas, initial delay of 120 and threshold of 3 and period of 15.

(120 * 2) + (3*15) + 120 = 405 seconds (6min 45sec)

Default: 60*1 + 3*10 + 120 = 210 (3,5 min)

Note: This ups the default from 3 minutes, to 3,5 minutes.