Build the project with GitHub Actions
The same MySQL test that is failing in Azure Pipelines is also failing
in GitHub Actions after this. Other than that everything should be green.

The docker commands used in the steps could very likely be improved
(is 'run' really needed very time or could 'exec' be used after starting
the container just once?), but this is a functional first step.