Add k8s and docker markers for dependency tests
This adds a two marks that can be used to mark tests that require kubernetes and/or docker.
There are more services that run during tests in Docker Compose, however, these services
are used only through the other tests that require Docker and Kubernetes, so we do not add
those marks (such as Helm) at this point in time and they would not have a use case.