Remove caching config for Trivy
The setup was removing caches before running and then adding caching which
looked at least like it did not make sense.

Moreover this also fixes an issue with the latest versions of Trivy where
image scanning would fail as the the command was interpreted as it was not
a `image` command but just the trivy command. This causing a failiure when
running image tests.